refactor: NoOp実装をnoop.rsモジュールに分離
lib.rsから約60行のNoOp実装をregistry/src/noop.rsに移動し可読性を向上
