+++ services/backend/server-core/registry/src/lib.rs
@@ -1,4 +1,5 @@
-use async_trait::async_trait;
+mod noop;
+
 use sea_orm::DatabaseConnection;
 use std::sync::Arc;
 use tracing;
@@ -294,65 +295,7 @@ pub fn create_repositories(db: DatabaseConnection) -> RepositoriesImpl {
     }
 }
 
-// --- NoOp implementations for local development ---
-
-struct NoOpLlmService;
-#[async_trait]
-impl LlmServiceTrait for NoOpLlmService {
-    async fn ask_question(
-        &self,
-        _question: domain::services::llm::Question,
-    ) -> Result<domain::services::llm::Answer, anyhow::Error> {
-        Err(anyhow::anyhow!("LlmService is not configured"))
-    }
-}
-
-struct NoOpDspFetcherService;
-#[async_trait]
-impl DspFetcherServiceTrait for NoOpDspFetcherService {
-    async fn fetch_dsps_data(
-        &self,
-        _start_date: Option<String>,
-        _end_date: String,
-    ) -> Result<Vec<application::services::dsp_fetcher::DspsData>, anyhow::Error> {
-        Err(anyhow::anyhow!("DspFetcherService is not configured"))
-    }
-    async fn fetch_gender_gen_data(
-        &self,
-        _date: String,
-    ) -> Result<Vec<application::services::dsp_fetcher::GenderGenData>, anyhow::Error> {
-        Err(anyhow::anyhow!("DspFetcherService is not configured"))
-    }
-    async fn fetch_sparse_data(
-        &self,
-        _date: String,
-    ) -> Result<Vec<application::services::dsp_fetcher::SparseData>, anyhow::Error> {
-        Err(anyhow::anyhow!("DspFetcherService is not configured"))
-    }
-}
-
-struct NoOpPushNotificationService;
-#[async_trait]
-impl PushNotificationServiceTrait for NoOpPushNotificationService {
-    async fn send_push_notification(
-        &self,
-        _notification: domain::services::notification::PushNotification,
-    ) -> Result<String, anyhow::Error> {
-        tracing::warn!("PushNotificationService is not configured, skipping notification");
-        Ok("noop".to_string())
-    }
-}
-
-struct NoOpOnchainFetcherService;
-#[async_trait]
-impl OnchainFetcherServiceTrait for NoOpOnchainFetcherService {
-    async fn fetch_credential_balances(
-        &self,
-        _evm_addrs: Vec<String>,
-    ) -> Result<Vec<application::services::onchain_fetcher::CredentialBalance>, anyhow::Error> {
-        Err(anyhow::anyhow!("OnchainFetcherService is not configured"))
-    }
-}
+use noop::{NoOpDspFetcherService, NoOpLlmService, NoOpOnchainFetcherService, NoOpPushNotificationService};
 
 pub async fn create_services() -> ServicesImpl {
     tracing::info!("Setup Services...");
